build_config/build_requirements.txt
# Build dependencies
pyinstaller>=6.3.0
pillow>=10.0.0  # For icon conversion